refactor linkchecer to refill threads every 24 hours
