+++ linkChecker/src/main/java/eu/clarin/curation/linkchecker/helpers/Configuration.java
@@ -19,14 +19,14 @@ public class Configuration {
     private static Properties properties = new Properties();
     private final static Logger _logger = LoggerFactory.getLogger(Configuration.class);
 
-    public static String DATABASE_NAME;
-    public static String DATABASE_URI;
     public static int TIMEOUT;
     public static int REDIRECT_FOLLOW_LIMIT;
     public static String USERAGENT;
     public static long CRAWLDELAY;
     public static boolean ONLY_BROKEN;
     public static Map<String, Long> CRAWLDELAYMAP = new HashMap<>();
+
+    public static MongoDatabase DATABASE;
 
     public static void loadConfigVariables(String configPath) {
         try {
@@ -35,13 +35,10 @@ public class Configuration {
 
             _logger.error("Can't load properties file: " + e.getMessage());
             System.exit(1);
-
         }
 
-        DATABASE_NAME = properties.getProperty("DATABASE_NAME");
         TIMEOUT = Integer.parseInt(properties.getProperty("TIMEOUT"));
         REDIRECT_FOLLOW_LIMIT = Integer.parseInt(properties.getProperty("REDIRECT_FOLLOW_LIMIT"));
-        DATABASE_URI = properties.getProperty("DATABASE_URI");
         USERAGENT = properties.getProperty("USERAGENT");
         CRAWLDELAY = Long.parseLong(properties.getProperty("CRAWLDELAY"));
         ONLY_BROKEN = Boolean.parseBoolean(properties.getProperty("ONLY_BROKEN"));
@@ -52,6 +49,26 @@ public class Configuration {
             long delay = Long.parseLong(crawlDelayEntry.split("=")[1]);
             CRAWLDELAYMAP.put(collection, delay);
         }
+
+        connectDatabase(properties.getProperty("DATABASE_NAME"),properties.getProperty("DATABASE_URI"));
+
+    }
+
+    private static void connectDatabase(String databaseName, String databaseURI) {
+        _logger.info("Connecting to database...");
+
+        MongoClient mongoClient;
+        if (databaseURI.isEmpty()) {//if it is empty, try localhost
+            mongoClient = MongoClients.create();
+        } else {
+            mongoClient = MongoClients.create(databaseURI);
+        }
+
+        MongoDatabase database = mongoClient.getDatabase(databaseName);
+
+        _logger.info("Connected to database.");
+
+        DATABASE = database;
 
     }
